+++ Python Module 1/ex3/ft_plant_factory.py
@@ -1,5 +1,5 @@
 class Plant:
-    def __init__(self, name: str, height: int,
+    def __init__(self, name: str, height: float,
                  age: int, growth: float) -> None:
         self.name = name
         self.height = height
